requirements.txt
discord.py[voice]
python-dotenv
requests
youtube_dl
selenium
bs4
lxml